Customer_support_back-end/requirements.txt
alembic==1.8.1
bidict==0.23.1
blinker==1.8.2
click==8.1.8
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.10.0
Flask-Migrate==3.1.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3.post0
h11==0.16.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
psycopg2-binary==2.9.5
PyJWT==2.9.0
python-dotenv==1.0.1
python-engineio==4.12.0
python-http-client==3.3.7
python-socketio==5.4.0
sendgrid==6.9.7
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==1.4.36
starkbank-ecdsa==2.2.0
Werkzeug==2.3.0
wsproto==1.2.0
zipp==3.20.2
